+++ news.md
@@ -1,5 +1,9 @@
 # retl 0.1.2.9000
 
+* `bqRefreshPartitionData()`, `bqTransformPartition()` - priority parameter added to the functions (#86).
+
+* `createRangeTable()` this function was fully replaced by `bqCreatePartitionTable()` and `bqTransformPartition()` functions (#86).
+
 * `bqCreatePartitionTable()` - added priority parameter that allows to execute biquery jobs in BATCH mode.
 
 * `dcPredictionBody()` - function is vectorised, which means that it can turn multiple transactions